move new plugin endpoints to apps
